Change type of NetworkHashPS to double
The type of NetworkHashPS was long and this resulted in overflow
at block 591855. The value was 6.988e+19 and
int64.max ~= 9e18. This commit changes the type from long to
double. Local testing against my own node has been performed and
this change fixed the problem.

+++ src/BitcoinLib/Responses/GetMiningInfoResponse.cs
@@ -11,11 +11,11 @@ namespace BitcoinLib.Responses
         public double Difficulty { get; set; }
         public string Errors { get; set; }
         public int GenProcLimit { get; set; }
-        public long NetworkHashPS { get; set; }
+        public double NetworkHashPS { get; set; }
         public int PooledTx { get; set; }
         public bool Testnet { get; set; }
         public string Chain { get; set; }
         public bool Generate { get; set; }
-        public long HashesPerSec { get; set; }
+        public double HashesPerSec { get; set; }
     }
 }
